Revert "items framework in place, need to create individual items now"
This reverts commit 22176acb9af621f72e18dcc5c8d250454878757e.
