--- .vscode/launch.json
@@ -1,10 +1,0 @@
-{
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": "Attach to Unity",
-            "type": "vstuc",
-            "request": "attach"
-        }
-     ]
-}
