--- Assets/Items.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 22f19c05fc243884ba9adaf23dc1675d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
