--- Assets/Items/Item.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bd17ef26097914e45a77ace2c892383b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
