--- Assets/Materials/HeldItem.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c0ea5a036ea5fb3429b5818d15b4dcf9
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
