--- Assets/Prefabs/Held Item.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a440c521f8ee14640882dc5257d41f8c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
