--- Assets/Prefabs/Item Image.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9bfcdb32057f2f54aa5410fb7d5b000a
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
