--- Assets/Scripts/Kart/HeldItem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3b67c9a93f4e93f418bee86824b97078
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
