--- Assets/Scripts/UI/ItemImage.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d09a55f6faf26df4f894df2e2b05cdbe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
