--- Assets/Scripts/UI/ItemSlotAnimator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1be87775e4fd82649b0a6d29cb8a19a2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
