+++ Packages/manifest.json
@@ -2,7 +2,7 @@
   "dependencies": {
     "com.unity.feature.development": "1.0.1",
     "com.unity.ide.rider": "3.0.18",
-    "com.unity.ide.visualstudio": "2.0.22",
+    "com.unity.ide.visualstudio": "2.0.17",
     "com.unity.ide.vscode": "1.2.5",
     "com.unity.inputsystem": "1.4.4",
     "com.unity.test-framework": "1.1.31",
