+++ Packages/packages-lock.json
@@ -38,7 +38,7 @@
       "url": "https://packages.unity.com"
     },
     "com.unity.ide.visualstudio": {
-      "version": "2.0.22",
+      "version": "2.0.17",
       "depth": 0,
       "source": "registry",
       "dependencies": {
